Removed warning message from editor configuration dialog
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4001 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/editorconfigurationdlg.cpp
@@ -794,7 +794,7 @@ void EditorConfigurationDlg::OnAutoCompDelete(wxCommandEvent& event)
     {
         m_AutoCompMap.erase(it);
         lstKeyword->Delete(sel);
-        if (sel >= lstKeyword->GetCount())
+        if (sel >= (int)(lstKeyword->GetCount()))
             sel = lstKeyword->GetCount() - 1;
         lstKeyword->SetSelection(sel);
         if (sel != -1)
